api: Add `emailAddressVisibility` to `InitialSnapshot`
This is for backward compatibility to the Zulip server versions prior to
Zulip 7.0 (FL 163) where there was realm-level `email_address_visibility`
policy for which a user can see other user's real email address. Search
for "email_address_visibility" in https://zulip.com/api/register-queue.

This will become handy in the next commit(s) where the user's real email
is shown in it's profile page.

Note: This field is removed in Zulip 7.0 (FL 163) and replaced with:
* https://zulip.com/api/update-settings#parameter-email_address_visibility
* https://zulip.com/api/update-realm-user-settings-defaults#parameter-email_address_visibility