narrow: Use TopicName.isSameAs in TopicNarrow.containsMessage
TopicNarrow.containsMessage used to use raw equality operator (==)
for comparing topics, which was not ideal for how we treat two
topics the same.

This would result in not displaying a new message of topic "T" when
looking at a topic narrow of topic "t" (treated message topics
case-sensitively in topic narrow).